+++ src/module/api/program.py
@@ -2,12 +2,11 @@ import signal
 import logging
 import os
 
-from fastapi.exceptions import HTTPException
-
-
+from fastapi import HTTPException, status, Depends
 from fastapi import FastAPI
 
 from module.core import Program
+from module.security import get_current_user
 
 logger = logging.getLogger(__name__)
 program = Program()
@@ -25,7 +24,9 @@ async def shutdown():
 
 
 @router.get("/api/v1/restart", tags=["program"])
-async def restart():
+async def restart(current_user=Depends(get_current_user)):
+    if not current_user:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="invalid token")
     try:
         program.restart()
         return {"status": "ok"}
@@ -36,7 +37,9 @@ async def restart():
 
 
 @router.get("/api/v1/start", tags=["program"])
-async def start():
+async def start(current_user=Depends(get_current_user)):
+    if not current_user:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="invalid token")
     try:
         program.start()
         return {"status": "ok"}
@@ -47,13 +50,17 @@ async def start():
 
 
 @router.get("/api/v1/stop", tags=["program"])
-async def stop():
+async def stop(current_user=Depends(get_current_user)):
+    if not current_user:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="invalid token")
     program.stop()
     return {"status": "ok"}
 
 
 @router.get("/api/v1/status", tags=["program"])
-async def status():
+async def status(current_user=Depends(get_current_user)):
+    if not current_user:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="invalid token")
     if not program.is_running:
         return {"status": "stop"}
     else:
@@ -61,7 +68,9 @@ async def status():
 
 
 @router.get("/api/v1/shutdown", tags=["program"])
-async def shutdown_program():
+async def shutdown_program(current_user=Depends(get_current_user)):
+    if not current_user:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="invalid token")
     program.stop()
     logger.info("Shutting down program...")
     os.kill(os.getpid(), signal.SIGINT)
@@ -70,10 +79,14 @@ async def shutdown_program():
 
 # Check status
 @router.get("/api/v1/check/downloader", tags=["check"])
-async def check_downloader_status():
+async def check_downloader_status(current_user=Depends(get_current_user)):
+    if not current_user:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="invalid token")
     return program.check_downloader()
 
 
 @router.get("/api/v1/check/rss", tags=["check"])
-async def check_rss_status():
+async def check_rss_status(current_user=Depends(get_current_user)):
+    if not current_user:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="invalid token")
     return program.check_analyser()
